requirements.txt
allure-behave==2.8.18
attrs==25.3.0
behave==1.2.6
behave-html-formatter==0.9.10
configparser==7.2.0
greenlet==3.2.0
parse==1.20.2
parse_type==0.6.4
playwright==1.51.0
pluggy==1.5.0
pyee==12.1.1
six==1.17.0
typing_extensions==4.13.2
allure-python-commons==2.8.18